+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -13,11 +13,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.interface21.webmvc.servlet.View;
+import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.AsIsHandlerAdaptor;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.ToBeHandlerAdapter;
 
 public class DispatcherServlet extends HttpServlet {
 
@@ -35,13 +34,12 @@ public class DispatcherServlet extends HttpServlet {
         handlerMappings = List.of(
                 new ManualHandlerMapping(),
                 new AnnotationHandlerMapping(getClass().getPackageName()));
-        handlerAdapters = List.of(new AsIsHandlerAdaptor(), new ToBeHandlerAdapter());
+        handlerAdapters = List.of(new ManualHandlerAdaptor(), new AnnotationHandlerAdapter());
         handlerMappings.forEach(HandlerMapping::initialize);
     }
 
     @Override
-    protected void service(final HttpServletRequest request, final HttpServletResponse response)
-            throws ServletException {
+    protected void service(final HttpServletRequest request, final HttpServletResponse response) throws ServletException {
         log.debug("Method : {}, Request URI : {}", request.getMethod(), request.getRequestURI());
         try {
             final var handler = getHandler(request);
@@ -57,7 +55,7 @@ public class DispatcherServlet extends HttpServlet {
 
     private Object getHandler(final HttpServletRequest request) {
         return handlerMappings.stream()
-                .flatMap(handlerMapping -> Stream.ofNullable(handlerMapping.getHandler(request)) )
+                .flatMap(handlerMapping -> Stream.ofNullable(handlerMapping.getHandler(request)))
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException(
                         "No handler found for request URI: " + request.getRequestURI()));
