+++ app/src/main/java/com/techcourse/ManualHandlerAdaptor.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.techcourse;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -6,9 +6,10 @@ import jakarta.servlet.http.HttpServletResponse;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
 import com.interface21.webmvc.servlet.view.JspView;
 
-public class AsIsHandlerAdaptor implements HandlerAdapter {
+public class ManualHandlerAdaptor implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
